do not replace agent symlink if it is ok
The problem here was:

* You're sitting at a terminal on host A.
* You ssh to host B, which sees that you've got an SSH_AUTH_SOCK
* ...so it points ~/.ssh/agent to that file
* ...and then you ssh again from B back to B
* ...so it does the same thing, pointing the agent at the sub-ssh
     agent.
* ...but then you log out of the B-to-B ssh
* ...and the agent file is now pointing to a dead socket.

This came up in practice when doing a git update, on B, to a repo
hosted on B.  The solution is simple: if the agent is okay, just let
it be.

+++ .zshenv
@@ -1,6 +1,9 @@
 export LANG=en_US.UTF-8
 
-if [ ! -z "$SSH_AUTH_SOCK" ] && [ "$SSH_AUTH_SOCK" != "$HOME/.ssh/agent" ]; then
+if [ ! -z "$SSH_AUTH_SOCK" ] \
+  && [ ! -e "$HOME/.ssh/agent" ] \
+  && [ "$SSH_AUTH_SOCK" != "$HOME/.ssh/agent" ]; \
+then
   ln -snf "$SSH_AUTH_SOCK" "$HOME/.ssh/agent"
   export SSH_AUTH_SOCK="$HOME/.ssh/agent"
 fi
